Move combinations to state
- Move combinations to state
- Move entire combinations related vars into a custom hook
